Monotone-Parent: e356443889853160f98a31849aede88bc4bac6bd Monotone-Revision: cdceff0f0c850bdd8d0af588d8e4b75eb75634a5
Monotone-Author: wsourdeau@inverse.ca
Monotone-Date: 2010-12-30T14:50:48
Monotone-Branch: ca.inverse.sogo

+++ OpenChange/NSValue+MAPIStore.m
@@ -56,11 +56,11 @@
   return value;
 }
 
-- (uint64_t *) asDoubleInMemCtx: (void *) memCtx
+- (double *) asDoubleInMemCtx: (void *) memCtx
 {
-  uint64_t *value;
+  double *value;
 
-  value = talloc (memCtx, uint64_t);
+  value = talloc (memCtx, double);
   *value = [self doubleValue];
 
   return value;
